+++ src/gitstore/__init__.py
@@ -3,11 +3,11 @@ from .mirror import MirrorDiff, RefChange
 from .fs import FS, retry_write
 from .tree import WalkEntry
 from .exceptions import StaleSnapshotError
-from .copy import ChangeReport, ChangeAction, ChangeError, FileEntry, FileType
+from .copy import ChangeReport, ChangeAction, ChangeError, FileEntry, FileType, disk_glob
 
 __all__ = [
     "GitStore", "ReflogEntry", "MirrorDiff", "RefChange",
     "FS", "retry_write", "StaleSnapshotError",
-    "ChangeReport", "ChangeAction", "ChangeError", "FileEntry", "FileType",
+    "ChangeReport", "ChangeAction", "ChangeError", "FileEntry", "FileType", "disk_glob",
     "WalkEntry",
 ]